+++ src/Editor/CMakeLists.txt
@@ -2,7 +2,7 @@ list(APPEND INC
 	"Action.h"
 	"Clipboard.h"
 	"Common.h"
-	"ConvertToOgg.h"
+	"ConvertAudio.h"
 	"Editing.h"
 	"Editor.h"
 	"FindOnsets.h"
@@ -29,7 +29,7 @@ list(APPEND SRC
 	"Action.cpp"
 	"Clipboard.cpp"
 	"Common.cpp"
-	"ConvertToOgg.cpp"
+	"ConvertAudio.cpp"
 	"Editing.cpp"
 	"Editor.cpp"
 	"FindOnsets.cpp"
