+++ src/Editor/Music.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <Editor/Sound.h>
-#include <Editor/ConvertToOgg.h>
+#include <Editor/ConvertAudio.h>
 
 namespace Vortex {
